requirements.txt
requests==2.31.0
python-dotenv==1.0.0
woocommerce==3.0.0
setuptools==68.2.2
mysql-connector-python==8.0.33 